3-tier-project/requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
cryptography==42.0.1